added a gemini rest api call
