+++ backend/server.js
@@ -14,6 +14,12 @@ import Blog from './Schema/Blog.js';
 import aws from 'aws-sdk';
 import Notification from './Schema/Notification.js';
 import Comment from './Schema/Comment.js';
+import fs from 'fs';
+import {
+  GoogleGenerativeAI,
+  HarmCategory,
+  HarmBlockThreshold,
+} from '@google/generative-ai';
 
 const server = express();
 let PORT = 3000;
@@ -1151,7 +1157,7 @@ server.post('/add-verification-token', (req, res) => {
     });
 });
 
-server.post('/check-verification-token', (req, res) => {
+server.post('/check-verification-token', async (req, res) => {
   let { verificationToken } = req.body;
 
   console.log(verificationToken);
@@ -1167,6 +1173,56 @@ server.post('/check-verification-token', (req, res) => {
     .catch((err) => {
       return res.status(500).json({ error: err.message });
     });
+});
+
+server.post('/build-travel-plan', async (req, res) => {
+  let { travelTo, travelFrom, travelWith, fromDate, toDate, specificActivity } =
+    req.body;
+  const MODEL_NAME = 'gemini-1.5-pro-latest';
+  const genAI = new GoogleGenerativeAI(process.env.GEMINI_API_KEY);
+  const model = genAI.getGenerativeModel({ model: MODEL_NAME });
+
+  const generationConfig = {
+    temperature: 1,
+    topK: 0,
+    topP: 0.95,
+    maxOutputTokens: 8192,
+  };
+
+  const safetySettings = [
+    {
+      category: HarmCategory.HARM_CATEGORY_HARASSMENT,
+      threshold: HarmBlockThreshold.BLOCK_MEDIUM_AND_ABOVE,
+    },
+    {
+      category: HarmCategory.HARM_CATEGORY_HATE_SPEECH,
+      threshold: HarmBlockThreshold.BLOCK_MEDIUM_AND_ABOVE,
+    },
+    {
+      category: HarmCategory.HARM_CATEGORY_SEXUALLY_EXPLICIT,
+      threshold: HarmBlockThreshold.BLOCK_MEDIUM_AND_ABOVE,
+    },
+    {
+      category: HarmCategory.HARM_CATEGORY_DANGEROUS_CONTENT,
+      threshold: HarmBlockThreshold.BLOCK_MEDIUM_AND_ABOVE,
+    },
+  ];
+
+  const chat = model.startChat({
+    generationConfig,
+    safetySettings,
+    history: [],
+  });
+
+  const context_prompt =
+    'User wants to build a travel plan. You are helpimg the user with provided information. Your response is going to coverted to json file. The query includes request_promt which is the form that the final travel plan should follow.';
+  const request_prompt = '';
+  const example_prompt = '';
+  const query = `${context_prompt}\nUser's preference: ${travelTo}, ${travelFrom}, ${travelWith}, ${fromDate} to ${toDate}, ${specificActivity}\nrequest_prompt: ${request_prompt}\nexample_prompt: ${example_prompt}`;
+
+  const result = await chat.sendMessage(query);
+  // return res.status(200).json(JSON.parse(result));
+  return res.status(200).json(result.response.candidates[0].content.parts[0].text);
 });
 
 server.listen(PORT, () => {
